GtkMinimal sample: Rework to be more tailored to a gtk builder sample
