+++ Samples/Gtk3/Builder/DemoWindow.cs
@@ -31,23 +31,12 @@ namespace GtkDemo
 
             var label = new Label("Gtk and C# - Very exciting isn't it?");
             notebook.InsertPage("Label", label, 1);
-
-            var button = new Button("Open!");
-            notebook.InsertPage("Dialogue", button, 2);
-            button.OnClicked += OpenDialog;
         }
 
         private void Button_Clicked(Button sender, EventArgs args)
         {
             Console.WriteLine("Hello World!");
             Resizable = !Resizable;
-        }
-
-        private void OpenDialog(Button sender, EventArgs args)
-        {
-            // TODO: Investigate adding widgets to GDialog
-            var dialog = new Dialog();
-            dialog.Run();
         }
     }
 }
