+++ Samples/Gtk3/Builder/data/demo_window.glade
@@ -4,7 +4,7 @@
   <requires lib="gtk+" version="3.18"/>
   <object class="GtkApplicationWindow" id="root">
     <property name="can_focus">False</property>
-    <property name="title" translatable="yes">Gtk C# Minimal Example</property>
+    <property name="title" translatable="yes">Gtk C# builder api example</property>
     <child>
       <object class="GtkBox" id="Box">
         <property name="visible">True</property>
