+++ Samples/Gtk3/Builder/data/gtk.png
